diffuser/datasets/minari-dataset-generation-scripts-main/scripts/antmaze/requirements.txt
numpy==1.26.4
scipy==1.12.0
packaging==24.0
gymnasium-robotics==1.2.4
minari==0.4.3
stable_baselines3@git+https://github.com/DLR-RM/stable-baselines3.git@f56ddeda10b1e3669a77a1c28c56944036286833
tqdm==4.66.2
